[5153] Addresses @kaikreuzer's review comment to use linuxsysfs instead of linux.sysfs in the bundle names
