--- bundles/config/org.eclipse.smarthome.config.discovery.usbserial.linux.sysfs/META-INF/MANIFEST.MF
@@ -1,16 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-Name: Eclipse SmartHome Configuration USB-Serial Discovery Linux sysfs Scanning
-Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Bundle-SymbolicName: org.eclipse.smarthome.config.discovery.usbserial.linux.sysfs
-Bundle-Vendor: Eclipse.org/SmartHome
-Bundle-Version: 0.10.0.qualifier
-Import-Package: 
- com.google.common.collect,
- org.eclipse.jdt.annotation;resolution:=optional,
- org.eclipse.smarthome.config.discovery.usbserial,
- org.eclipse.smarthome.core.common,
- org.osgi.framework,
- org.osgi.service.component,
- org.slf4j
-Service-Component: OSGI-INF/*.xml
