--- bundles/config/org.eclipse.smarthome.config.discovery.usbserial.linux.sysfs/pom.xml
@@ -1,17 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-
-  <parent>
-    <groupId>org.eclipse.smarthome.bundles</groupId>
-    <artifactId>config</artifactId>
-    <version>0.10.0-SNAPSHOT</version>
-  </parent>
-
-  <modelVersion>4.0.0</modelVersion>
-  <groupId>org.eclipse.smarthome.config</groupId>
-  <artifactId>org.eclipse.smarthome.config.discovery.usbserial.linux.sysfs</artifactId>
-
-  <name>Eclipse SmartHome Configuration USB-Serial Discovery for Linux using sysfs scanning</name>
-
-  <packaging>eclipse-plugin</packaging>
-
-</project>
