+++ bundles/config/org.eclipse.smarthome.config.discovery.usbserial.linuxsysfs.test/.classpath
