+++ bundles/config/org.eclipse.smarthome.config.discovery.usbserial.linuxsysfs.test/.settings/org.eclipse.jdt.core.prefs
