+++ bundles/config/org.eclipse.smarthome.config.discovery.usbserial.linuxsysfs.test/META-INF/MANIFEST.MF
@@ -1,0 +1,22 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: Eclipse SmartHome Configuration Discovery Usb-Serial Linux Sysfs Tests
+Bundle-RequiredExecutionEnvironment: JavaSE-1.8
+Bundle-SymbolicName: org.eclipse.smarthome.config.discovery.usbserial.linuxsysfs.test;singleton:=true
+Bundle-Vendor: Eclipse.org/SmartHome
+Bundle-Version: 0.10.0.qualifier
+Fragment-Host: org.eclipse.smarthome.config.discovery.usbserial.linuxsysfs
+Import-Package:  
+ com.google.common.collect,
+ org.eclipse.jdt.annotation;resolution:=optional,
+ org.eclipse.smarthome.config.discovery.usbserial.testutil,
+ org.eclipse.smarthome.test.java,
+ org.hamcrest;core=split,
+ org.hamcrest.collection,
+ org.hamcrest.core,
+ org.junit;version="4.0.0",
+ org.junit.rules,
+ org.mockito,
+ org.mockito.invocation,
+ org.mockito.stubbing,
+ org.mockito.verification
