+++ bundles/config/org.eclipse.smarthome.config.discovery.usbserial.linuxsysfs/.settings/org.eclipse.jdt.core.prefs
