+++ bundles/config/org.eclipse.smarthome.config.discovery.usbserial.linuxsysfs/OSGI-INF/.gitignore
